+++ zipapps/__main__.py
@@ -193,6 +193,16 @@ def main():
         help='Add the ensurepip package to your pyz file, works for '
         'embed-python(windows) or other python versions without `pip`'
         ' installed but `lazy-install` mode is enabled. [EXPERIMENTAL]')
+    parser.add_argument(
+        '--layer-mode',
+        action='store_true',
+        dest='layer_mode',
+        help='Layer mode for the serverless use case, __main__.py / ensure_zipapps.py / activate_zipapps.py files will not be set in this mode.')
+    parser.add_argument(
+        '--layer-mode-prefix',
+        default='python',
+        dest='layer_mode_prefix',
+        help='Layer mode for the serverless use case, __main__.py / ensure_zipapps.py / activate_zipapps.py files will not be set in this mode.')
     if len(sys.argv) == 1:
         return parser.print_help()
     args, pip_args = parser.parse_known_args()
@@ -221,6 +231,8 @@ def main():
         sys_paths=args.sys_paths,
         python_version_slice=int(args.python_version_slice),
         ensure_pip=args.ensure_pip,
+        layer_mode=args.layer_mode,
+        layer_mode_prefix=args.layer_mode_prefix,
     )