+++ changelog.md
@@ -1,6 +1,12 @@
 
 # Changelogs
 
+- 2021.09.21
+  - `layer-mode` for serverless layers
+    - https://docs.aws.amazon.com/lambda/latest/dg/configuration-layers.html
+    - this mode will not activate the sys.path of this zip file.
+    - `python3 -m zipapps --layer-mode --layer-mode-prefix=python -o layer.zip -r requirements.txt`
+      - maybe add `-cc` to generate the `.pyc` files
 - 2021.05.17
   - `lazy-install` mode will re-install packages while pip args including `-U` or `--upgrade`.
 - 2021.04.29